requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
spacy>=3.4.0
sentence-transformers>=2.2.0
rapidfuzz>=2.0.0
python-dateutil>=2.8.0
reportlab>=3.6.0
openpyxl>=3.0.0
plotly>=5.17.0